+++ src/main/java/resources/OrdemDeServicoResource.java
@@ -9,15 +9,12 @@ import domain.Cliente;
 import domain.OrdemDeServico;
 import java.time.LocalDateTime;
 import javax.inject.Inject;
-import javax.ws.rs.Consumes;
 import javax.ws.rs.DELETE;
 import javax.ws.rs.GET;
 import javax.ws.rs.POST;
 import javax.ws.rs.PUT;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
-import javax.ws.rs.Produces;
-import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 import services.OrdemDeServicoService;
 
@@ -34,7 +31,7 @@ public class OrdemDeServicoResource {
 
     @GET
     @Path("{id}")
-    public Response getTodo(@PathParam("id") Long id) {
+    public Response getTodo(@PathParam("id") Integer id) {
         OrdemDeServico ordemDeServico = ordemDeServicoService.findById(id);
 
         return Response.ok(ordemDeServico).build();
@@ -48,7 +45,7 @@ public class OrdemDeServicoResource {
 
     @PUT
     @Path("{id}")
-    public Response update(@PathParam("id") Long id, Cliente cliente) {
+    public Response update(@PathParam("id") Integer id, Cliente cliente) {
         OrdemDeServico OrdemDeServicoUpdate = ordemDeServicoService.findById(id);
         OrdemDeServicoUpdate.setInstante(LocalDateTime.MIN);
         ordemDeServicoService.update(OrdemDeServicoUpdate);
@@ -58,7 +55,7 @@ public class OrdemDeServicoResource {
 
     @DELETE
     @Path("{id}")
-    public Response delete(@PathParam("id") Long id) {
+    public Response delete(@PathParam("id") Integer id) {
         OrdemDeServico ordemDeServico = ordemDeServicoService.findById(id);
 
         ordemDeServicoService.delete(ordemDeServico);
